+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [9.23.20](https://github.com/dereekb/dbx-components/compare/v9.23.19-dev...v9.23.20) (2023-05-19)
+
+
+### Bug Fixes
+
+* calendar selection store min/max range with filter fix ([865ef18](https://github.com/dereekb/dbx-components/commit/865ef18d2456a6be4f040c485196a3d6eef2386b))
+* fixed calendar selection end being before start when using a filter ([25f905f](https://github.com/dereekb/dbx-components/commit/25f905f6da092e4b2fbc42fdacd2dfb9d9c7eff1))
+* fixed markerClasses usage in DbxMapboxMarkerComponent ([d0a0b18](https://github.com/dereekb/dbx-components/commit/d0a0b1832cfc621627f0ca60f91251fea6f0aa92))
+
+
+
 ## [9.23.19](https://github.com/dereekb/dbx-components/compare/v9.23.18-dev...v9.23.19) (2023-05-11)